feat: rename second form to second phase
