+++ IsekaiMod/Config/Blueprints.json
@@ -526,10 +526,9 @@
     "CorruptAuraAreaBuff": "44a2f1f6-ec06-4f3d-8a41-c5debba18ea6",
     "CorruptAuraAbility": "1ac5a96b-4861-4b77-8314-f2684d48a45f",
     "CorruptAuraFeature": "3852c32f-e3ac-45dc-bb14-d868a87467ac",
-    "SecondFormBuffEffect": "1204ffd1-2371-4497-bbde-e839ee202908",
-    "SecondFormBuffTrigger": "7a626a0a-758a-4584-a90c-cb6032014bb7",
-    "SecondFormBuff": "2601c5aa-4f9b-450f-8a50-69f36d694d3d",
-    "SecondFormFeature": "56365f1f-cf02-4ad7-873f-5b6dd6becbb2",
+    "SecondPhaseBuffEffect": "1204ffd1-2371-4497-bbde-e839ee202908",
+    "SecondPhaseBuff": "2601c5aa-4f9b-450f-8a50-69f36d694d3d",
+    "SecondPhaseFeature": "56365f1f-cf02-4ad7-873f-5b6dd6becbb2",
 
     // Patch Prestige Class Spellbook
     "LoremasterIsekai": "4133ad57-ec5b-4562-96d3-613e90418597",
