+++ IsekaiMod/IsekaiMod.csproj
@@ -342,7 +342,7 @@
     <Compile Include="Content\Features\IsekaiProtagonist\Archetypes\Overlord\CorruptAuraFeature.cs" />
     <Compile Include="Content\Features\IsekaiProtagonist\Archetypes\Overlord\OverlordLegacySelection.cs" />
     <Compile Include="Content\Features\IsekaiProtagonist\Archetypes\Overlord\OverlordProficiencies.cs" />
-    <Compile Include="Content\Features\IsekaiProtagonist\Archetypes\Overlord\SecondFormFeature.cs" />
+    <Compile Include="Content\Features\IsekaiProtagonist\Archetypes\Overlord\SecondPhaseFeature.cs" />
     <Compile Include="Content\Features\IsekaiProtagonist\OverpoweredAbility\TabletopTweaksBase\AutoBurning.cs" />
     <Compile Include="Content\Features\IsekaiProtagonist\OverpoweredAbility\TabletopTweaksBase\AutoElemental.cs" />
     <Compile Include="Content\Features\IsekaiProtagonist\OverpoweredAbility\TabletopTweaksBase\AutoEncouraging.cs" />
